+++ crates/pessimistic-proof/Cargo.toml
@@ -27,4 +27,5 @@ rand = "0.8.5"
 sp1-sdk = { git = "https://github.com/succinctlabs/sp1" }
 ethers-signers = "2.0.14"
 sp1-zkvm = { git = "https://github.com/succinctlabs/sp1" }
+sp1-core = { git = "https://github.com/succinctlabs/sp1" }
