+++ crates/pessimistic-proof/tests/e2e_local_pp.rs
@@ -1,57 +1,22 @@
-use std::collections::{BTreeMap, BTreeSet};
-
-use ethers_signers::{LocalWallet, Signer};
-use lazy_static::lazy_static;
 use pessimistic_proof::{
-    bridge_exit::{commit_bridge_exits, BridgeExit, LeafType, NetworkId, TokenInfo},
-    imported_bridge_exit::{commit_imported_bridge_exits, ImportedBridgeExit},
-    keccak::{keccak256_combine, Digest},
-    local_balance_tree::{LocalBalancePath, LocalBalanceTree, LOCAL_BALANCE_TREE_DEPTH},
-    local_exit_tree::{data::LocalExitTreeData, hasher::Keccak256Hasher},
-    multi_batch_header::MultiBatchHeader,
-    nullifier_tree::{FromBool, NullifierKey, NullifierPath, NullifierTree, NULLIFIER_TREE_DEPTH},
-    utils::smt::Smt,
-    LocalNetworkState,
     bridge_exit::TokenInfo, local_balance_tree::LocalBalanceTree,
     multi_batch_header::MultiBatchHeader, nullifier_tree::NullifierTree, LocalNetworkState,
 };
-use rand::{random, thread_rng};
-use reth_primitives::{address, Address, Signature, U256};
 use rand::random;
 use reth_primitives::U256;
 use sp1_sdk::{utils, ProverClient, SP1Stdin};
-
-mod test_utils;
-
 use test_utils::{
     forest::Forest,
     sample_data::{ETH, NETWORK_A, NETWORK_B, USDC},
     PESSIMISTIC_PROOF_ELF,
 };
 
+use crate::test_utils::forest::signing_utils;
+
+mod test_utils;
+
 fn u(x: u64) -> U256 {
     x.try_into().unwrap()
-}
-
-fn signing_utils(
-    imported_bridge_exits: &[(ImportedBridgeExit, NullifierPath<Keccak256Hasher>)],
-    bridge_exits: &[BridgeExit],
-) -> (Digest, Address, Signature) {
-    let imported_hash =
-        commit_imported_bridge_exits(imported_bridge_exits.iter().map(|(exit, _)| exit));
-    let exit_hash = commit_bridge_exits(bridge_exits.iter());
-    let combined_hash = keccak256_combine([exit_hash.as_slice(), imported_hash.as_slice()]);
-
-    let wallet = LocalWallet::new(&mut thread_rng());
-    let signer = wallet.address();
-    let signature = wallet.sign_hash(combined_hash.into()).unwrap();
-    let signature = Signature {
-        r: U256::from_limbs(signature.r.0),
-        s: U256::from_limbs(signature.s.0),
-        odd_y_parity: signature.recovery_id().unwrap().is_y_odd(),
-    };
-
-    (imported_hash, signer.0.into(), signature)
 }
 
 fn e2e_local_pp_simple_helper(
